test(useFetchProducts): describe 별 테스트 분리
